requirements.txt
huggingface_hub
torch
pymysql
streamlit
pandas
numpy
python-dotenv
pywhatkit